add redo button when presenting trials

+++ src/trial_manager.ts
@@ -1,4 +1,4 @@
-import { WebGLRenderer } from "three";
+import { Scene, WebGLRenderer } from "three";
 import { startAndStreamHandDataToMain } from "./hand_capture";
 import { clearDisplayIndefinitely, displayString, displayStringIndefinitely, font, loadFont } from "./text_display";
 import { completeTrial, getDemonstration } from "./http_handler";
@@ -53,21 +53,45 @@ export async function performTrial(
         trialToPerform.instructions,
         "Place your hands inside the box when ready",
         "Remove your hands to continue",
-        scene);
+        scene).completion;
 
     const demonstration = new GestureDemonstration(trialToPerform.trial_name);
 
-    for (let i = 0; i < trialToPerform.gestures.length; i++) {
-        const gestureToPerform = trialToPerform.gestures[i];
-        await performGesture(
-            gestureToPerform,
-            getGestureLocator(i),
-            scene,
-            renderer,
-            demonstration
-        )
-    }
+    // So the first gesture does not present a 'redo' option
+    let offerRedo = false;
 
+    for (let askGestureIndex = 0; askGestureIndex < trialToPerform.gestures.length;) {
+        const gestureToPerform = trialToPerform.gestures[askGestureIndex];
+        
+        console.log('%ctrial_manager.ts line:64 i', 'color: #007acc;', askGestureIndex);
+        console.log('%ctrial_manager.ts line:68 offerRedo', 'color: #007acc;', offerRedo);
+
+        const record = await displayGestureInstructions(
+            gestureToPerform,
+            getGestureLocator(askGestureIndex),
+            scene,
+            demonstration,
+            offerRedo  
+        );
+
+        if (record) {
+            await performGesture(
+                gestureToPerform,
+                getGestureLocator(askGestureIndex),
+                scene,
+                renderer,
+            );
+
+            offerRedo = true;
+            askGestureIndex++;
+        } else {
+
+            offerRedo = false;
+            askGestureIndex--;
+        } 
+        
+    }
+    
     function getGestureLocator(gesture_index: number): GestureLocator {
         return {
             project_name,
@@ -86,15 +110,18 @@ async function startDemonstrationIfExists(project_name: string, gesture_id: stri
     const demonstrationData = await getDemonstration(project_name, gesture_id);
 
     if (demonstrationData) {
-        console.log('%ctrial_manager.ts line:72 demonstrationData', 'color: #007acc;', demonstrationData);
         demonstration.load(demonstrationData)
         demonstration.startPlaybackLoop();
         console.log("demonstration started");
     }
 }
 
-async function performGesture(gesture: Gesture, gestureLocator: GestureLocator, scene: THREE.Scene, renderer: WebGLRenderer, demonstration: GestureDemonstration) {
-
+async function displayGestureInstructions(
+    gesture: Gesture, 
+    gestureLocator: GestureLocator, 
+    scene: Scene, 
+    demonstration: GestureDemonstration,
+    offerRedo = false) {
     startDemonstrationIfExists(gestureLocator.project_name, gesture.gesture_id, demonstration);
 
     const instructions = displaySkipableInstruction(
@@ -103,35 +130,38 @@ async function performGesture(gesture: Gesture, gestureLocator: GestureLocator, 
         "Remove your hands to start recording",
         scene, "next");
 
-    const undo = createUndoButton("undo");
+    let result: string;
 
-    const result = await Promise.any([
-        instructions.completion,
-        undo.completion
-    ])
+    if (offerRedo) {
+        const undo = createUndoButton("undo");
+    
+        result = await Promise.any([
+            instructions.completion,
+            undo.completion
+        ]);
+    
+        console.log('%ctrial_manager.ts line:143 result', 'color: #007acc;', result);
 
-    demonstration.stopPlayback();
-
-    switch (result) {
-        case "next":
-            undo.delete();
-
-            const durationMs = gesture.duration * 1000;
-            await Promise.all([
-                startAndStreamHandDataToMain(durationMs, renderer, gestureLocator),
-                displayString(`recording gesture for ${gesture.duration}s`, durationMs, scene)
-            ]);
-            break;
-        case "undo":
-            instructions.delete();
-
-            await performGesture(
-                gesture,
-                gestureLocator,
-                scene,
-                renderer,
-                demonstration
-            );
+        switch (result) {
+            case "next":
+                undo.delete();
+                break;
+            case "undo":
+                instructions.delete();
+                break;
+        };
+    } else {
+        result = await instructions.completion;
     }
 
+    demonstration.stopPlayback();
+    return result == "next";
+}
+
+async function performGesture(gesture: Gesture, gestureLocator: GestureLocator, scene: THREE.Scene, renderer: WebGLRenderer) {
+    const durationMs = gesture.duration * 1000;
+    await Promise.all([
+        startAndStreamHandDataToMain(durationMs, renderer, gestureLocator),
+        displayString(`recording ${gesture.gesture_name} for ${gesture.duration} seconds`, durationMs, scene)
+    ]);
 }
